fix issue with language class
