+++ app/src/main/java/com/vultisig/wallet/ui/models/settings/LanguageSettingViewModel.kt
@@ -8,6 +8,7 @@ import androidx.appcompat.app.AppCompatDelegate
 import androidx.core.os.LocaleListCompat
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.vultisig.wallet.data.models.Language
 import com.vultisig.wallet.data.models.settings.AppLanguage
 import com.vultisig.wallet.data.models.settings.AppLanguage.Companion.fromName
 import com.vultisig.wallet.data.repositories.AppLocaleRepository
