+++ data/src/main/kotlin/com/vultisig/wallet/data/models/Language.kt
@@ -1,0 +1,6 @@
+package com.vultisig.wallet.data.models
+
+data class Language(
+    val mainName: String = "",
+    val englishName: String? = null
+)
