+++ data/src/main/kotlin/com/vultisig/wallet/data/repositories/DataRepositoriesModule.kt
@@ -57,4 +57,10 @@ internal interface DataRepositoriesModule {
     fun bindTokenPriceRepository(
         impl: TokenPriceRepositoryImpl,
     ): TokenPriceRepository
+
+    @Binds
+    @Singleton
+    fun bindAppLocaleRepository(
+        impl: AppLocaleRepositoryImpl,
+    ): AppLocaleRepository
 }
